requirements.txt
openai>=1.0
tiktoken>=0.5
python-dotenv>=1.0
pydantic>=2.0
pytest>=7.0
